+++ L2J_Mobius_CT_0_Interlude/java/org/l2jmobius/gameserver/model/multisell/Entry.java
@@ -84,7 +84,7 @@ public class Entry
 		return _stackable;
 	}
 	
-	public long getTaxAmount()
+	public int getTaxAmount()
 	{
 		return 0;
 	}
